fixed a conversion issue
git-svn-id: http://pubsvn.inavare.net/openscada/modules/atlantis@2317 1b50e8ee-deed-0310-b1de-ab401e1c4d63

+++ core/common/src/org/openscada/core/Variant.java
@@ -222,7 +222,7 @@ public class Variant
             if ( _value instanceof Long )
                 return ((Long)_value).longValue ();
             if ( _value instanceof String )
-                return Integer.parseInt ((String)_value);
+                return Long.parseLong ((String)_value);
         }
         catch ( NumberFormatException e )
         {
